허용오차 적용 완료 및 전체 로직 수정

+++ mode2_test.py
@@ -101,7 +101,7 @@ class Mode_1:
                     break
 
                 try:
-                    rows = [int(x) for x in row]
+                    rows = [float(x) for x in row]
                     final_arr.append(rows)
 
                 except ValueError:
